+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for ADBS free DB version details
 - Support for MySQL Analytics added
+- Support for Private DNS Names added
 
 ## 4.2.0 (November 04, 2020)